Real file sytem test
