+++ packages/libs/datastore/src/file-system/file-system.ts
@@ -1,5 +1,3 @@
-import { ReadUri } from "@azure-tools/uri";
-
 /**
  * Virtual filesystem, to find and read files.
  */
@@ -28,35 +26,4 @@ export interface ILegacyFileSystem {
    * @deprecated, use @see {IFileSystem.read}
    */
   ReadFile(uri: string): Promise<string>;
-}
-
-const MAX_RETRY_COUNT = 3;
-/**
- * ReadUri with retries in the case of a remote url and retryable http error.
- * @param uri uri
- * @param headers optional headers for the request.
- */
-export async function readUriWithRetries(uri: string, headers: { [name: string]: string } = {}): Promise<string> {
-  let tryed = 1;
-  for (;;) {
-    try {
-      return await ReadUri(uri, headers);
-    } catch (e) {
-      tryed++;
-      if (isRetryableStatusCode(e.statusCode) && tryed <= MAX_RETRY_COUNT) {
-        // eslint-disable-next-line no-console
-        console.error(`Failed to load uri ${uri}, trying again (${tryed}/${MAX_RETRY_COUNT})`, e);
-      } else {
-        throw e;
-      }
-    }
-  }
-}
-
-/**
- * @param statusCode Error status code.
- * @returns Boolean if this request should be retried.
- */
-function isRetryableStatusCode(statusCode: number): boolean {
-  return statusCode >= 500 || statusCode === 0;
 }
