+++ packages/libs/datastore/src/file-system/real-file-system.ts
@@ -1,6 +1,7 @@
-import { EnumerateFiles, WriteString } from "@azure-tools/uri";
-import { IFileSystem, readUriWithRetries } from "./file-system";
+import { EnumerateFiles, ReadUri, WriteString } from "@azure-tools/uri";
+import { IFileSystem } from "./file-system";
 import * as Constants from "../constants";
+import { UriNotFoundError } from "./errors";
 
 export class RealFileSystem implements IFileSystem {
   public constructor() {}
@@ -9,15 +10,8 @@ export class RealFileSystem implements IFileSystem {
     return EnumerateFiles(folderUri, [Constants.DefaultConfiguration]);
   }
 
-  public async read(uri: string): Promise<string> {
-    try {
-      return await readUriWithRetries(uri);
-    } catch (e) {
-      console.error("E", Object.getOwnPropertyNames(e));
-      console.error(e.code);
-      console.error(e.constructor.name);
-      throw e;
-    }
+  public async read(uri: string, headers: { [name: string]: string } = {}): Promise<string> {
+    return await readUriWithRetries(uri, headers);
   }
 
   public async write(uri: string, content: string): Promise<void> {
@@ -37,4 +31,62 @@ export class RealFileSystem implements IFileSystem {
   public async EnumerateFileUris(folderUri: string): Promise<Array<string>> {
     return this.list(folderUri);
   }
+}
+
+const MAX_RETRY_COUNT = 3;
+/**
+ * ReadUri with retries in the case of a remote url and retryable http error.
+ * @param uri uri
+ * @param headers optional headers for the request.
+ */
+export async function readUriWithRetries(uri: string, headers: { [name: string]: string } = {}): Promise<string> {
+  let tryed = 1;
+  for (;;) {
+    try {
+      return await ReadUri(uri, headers);
+    } catch (e) {
+      tryed++;
+      if (isRetryableStatusCode(e.statusCode) && tryed <= MAX_RETRY_COUNT) {
+        // eslint-disable-next-line no-console
+        console.error(`Failed to load uri ${uri}, trying again (${tryed}/${MAX_RETRY_COUNT})`, e);
+      } else {
+        throw processError(uri, e);
+      }
+    }
+  }
+}
+
+export const KnownUriErrorCode = {
+  NotFound: "ENOTFOUND",
+};
+
+/**
+ * @param statusCode Error status code.
+ * @returns Boolean if this request should be retried.
+ */
+function isRetryableStatusCode(statusCode: number): boolean {
+  return statusCode >= 500 || statusCode === 0;
+}
+
+/**
+ * Convert external errors into known error types when able.
+ * @param uri Uri that was loaded.
+ * @param error Error.
+ * @returns new error if error is known or the same error otherwise.
+ */
+function processError(uri: string, error: Error) {
+  if (isCodeError(error)) {
+    if (error.code === KnownUriErrorCode.NotFound) {
+      return new UriNotFoundError(uri, `${uri} is not found: ${error.message}`);
+    }
+  }
+  return error;
+}
+
+interface CodeError extends Error {
+  code: string;
+}
+
+function isCodeError(error: unknown): error is CodeError {
+  return typeof error === "object" && error != null && "code" in error && (error as any)["code"] != null;
 }
